// root/src/modules/interfaces/enums/LensActionType.ts
export enum LensActionType {
    None = "none",
    CreateDiscussion = "createDiscussion",
    CreateSolution = "createSolution",
    CreateSubtreeLink = "createSubtreeLink",
    CreateSubtreeAndLink = "createSubtreeAndLink",
    ConvertBranchToSubtree = "convertBranchToSubtree",
    StashBranch = "stashBranch",
    MoveBranch = "moveBranch",
    MoveToHere = "moveToHere",
    ShowDetails = "showDetails",
    Refresh = "refresh",
    CloneBranch = "cloneBranch",
    CloneToHere = "cloneToHere",
    ShowSubtreeHub = "showSubtreeHub",
    ShowTreeEditor = "showTreeEditor",
    ViewBranches = "viewBranches",
    ViewProject = "viewProject",
    ShowBotHub = "showBotHub",
    ShowBotAliasHub = "showBotAliasHub",
    ShowBotDraftHub = "showBotDraftHub",
    CreateAlias = "createAlias",
    DeleteAlias = "deleteAlias",
    DeleteDraft = "deleteDraft",
    EditAlias = "editAlias",
    ShowTreeHub = "showTreeHub",
    CreateTree = "createTree",
    MarkSocket = "mapSocket",
}