// root/src/modules/interfaces/enums/StageTitle.ts
export enum StageTitle {

    None = "none",

    // Nodes
    LensEditNode = "lensEditNode",
    LensCreateNode = "lensCreateNode",
    LensEditHole = "lensEditHole",
    LensNodeAlts = "lensNodeAlts",

    // Search nodes
    LensSearchNodes = 'lensSearchNodes',

    // Hubs
    LensTreeHub = 'lensTreeHub',
    LensSubtreeHub = 'lensSubtreeHub',
    LensPublish = 'lensPublish',
    LensAliasHub = 'lensAliasHub',
    LensDraftHub = 'lensDraftHub',

    // Trees
    LensCreateTree = 'lensCreateTree',
    LensEditTree = 'lensEditTree',
    LensCloneTree = 'lensCloneTree',

    // Subtrees
    LensCreateSubtree = 'lensCreateSubtree',
    LensEditSubtree = 'lensEditSubtree',

    // Subtree links
    LensEditSubtreeLink = 'lensEditSubtreeLink',
    LensSwapSubtreeLink = 'lensSwapSubtreeLink',
    LensCreateSubtreeLink = 'lensCreateSubtreeLink',
    LensCreateSubtreeAndLink = 'lensCreateSubtreeAndLink',
    LensMarkSocket = 'lensMarkSocket',

    // Branch actions
    LensMoveBranch = 'lensMoveBranch',
    LensStashBranch = 'lensStashBranch',
    LensCloneBranch = 'lensCloneBranch',
    LensBranchToSubtree = 'lensBranchToSubtree',

    // Trees extra
    LensFilterTrees = 'lensFilterTrees',
    LensTreeHistory = 'lensTreeHistory',
    LensSubtreeShape = 'lensSubtreeShape',
    LensSubtreeSpread = 'lensSubtreeSpread',
    LensTreeTags = 'lensTreeTags',
    LensValidations = 'lensValidations',

    // Bots
    LensFilterBots = 'lensFilterBots',
    LensEditAlias = 'lensEditAlias',
    LensCreateAlias = 'lensCreateAlias',

    // Settings
    LensSettings = 'lensSettings',
    LensViewSettings = 'lensViewSettings',

    // Alerts
    LensNotification = 'lensNotification'
}