// root/src/modules/interfaces/enums/TabType.ts
export enum TabType {
    None = "none",
    Node = "node",
    Search = 'search',
    Filter = 'filter',
    Alts = 'alts',
    UserViews = 'userViews',
    Settings = 'settings',
    Alerts = 'alerts',
    Validation = 'validation',
    Spread = 'spread',
    Shape = 'shape',
    History = 'history',
    Tags = 'tags',
    Tree = 'tree',
    Bot = 'bot'
}